Wrong  test fixed
