+++ src/test/java/io/split/android/client/SplitClientImplTelemetryTest.java
@@ -1,14 +1,9 @@
 package io.split.android.client;
 
-import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyBoolean;
-import static org.mockito.ArgumentMatchers.anyDouble;
 import static org.mockito.ArgumentMatchers.anyList;
-import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.ArgumentMatchers.anyMap;
 import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
@@ -20,27 +15,6 @@ import java.util.Collections;
 import io.split.android.client.telemetry.model.Method;
 
 public class SplitClientImplTelemetryTest extends SplitClientImplBaseTest {
-
-    @Test
-    public void trackRecordsLatencyInEvaluationProducer() {
-        ProcessedEventProperties processedEventProperties = mock(ProcessedEventProperties.class);
-        when(processedEventProperties.isValid()).thenReturn(true);
-        when(eventsTracker.track(any(), any(), any(), anyDouble(), any())).thenReturn(true);
-        splitClient.track("any");
-
-        verify(telemetryStorageProducer).recordLatency(eq(Method.TRACK), anyLong());
-    }
-
-    @Test
-    public void trackRecordsExceptionInCaseThereIsOne() {
-        when(eventsTracker.track(any(), any(), any(), anyDouble(), any())).thenAnswer(invocation -> {
-            throw new Exception("test exception");
-        });
-
-        splitClient.track("event");
-
-        verify(telemetryStorageProducer).recordException(Method.TRACK);
-    }
 
     @Test
     public void getTreatmentRecordsException() {
